move formHelper functions to createDialog file
