+++ web/src/app/schedules/calendar-subscribe/CalendarSubscribeCreateDialog.js
@@ -4,8 +4,10 @@ import { useMutation } from '@apollo/react-hooks'
 import gql from 'graphql-tag'
 import FormDialog from '../../dialogs/FormDialog'
 import CalendarSubscribeForm from './CalendarSubscribeForm'
-import { getForm, FormTitle, getSubtitle } from './formHelper'
 import { fieldErrors, nonFieldErrors } from '../../util/errutil'
+import { makeStyles } from "@material-ui/core"
+import { CheckCircleOutline as SuccessIcon } from "@material-ui/icons"
+import CalenderSuccessForm from "./CalendarSuccessForm"
 
 const mutation = gql`
   mutation($input: CreateUserCalendarSubscriptionInput!) {
@@ -15,13 +17,38 @@ const mutation = gql`
   }
 `
 
+const useStyles = makeStyles(theme => ({
+  successIcon: {
+    marginRight: theme.spacing(1),
+  },
+  successTitle: {
+    color: 'green',
+    display: 'flex',
+    alignItems: 'center',
+  },
+}))
+
 const MOCK_URL =
   'www.calendarlabs.com/ical-calendar/ics/22/Chicago_Cubs_-_MLB.ics'
 
 const SUBTITLE =
   'Create a unique iCalendar subscription URL that can be used in your preferred calendar application.'
 
+export function getSubtitle(isComplete, defaultSubtitle) {
+  const completedSubtitle =
+    'Your subscription has been created! You can' +
+    ' manage your subscriptions from your profile at any time.'
+
+  return isComplete ? completedSubtitle : defaultSubtitle
+}
+
+export function getForm(isComplete, defaultForm, url) {
+  return isComplete ? <CalenderSuccessForm url={url} /> : defaultForm
+}
+
 export default function CalendarSubscribeCreateDialog(props) {
+  const classes = useStyles()
+
   const [value, setValue] = useState({
     name: '',
     scheduleID: props.scheduleID || null,
@@ -54,7 +81,12 @@ export default function CalendarSubscribeCreateDialog(props) {
 
   return (
     <FormDialog
-      title={FormTitle(isComplete, 'Create New Calendar Subscription')}
+      title={isComplete ? (
+        <div className={classes.successTitle}>
+          <SuccessIcon className={classes.successIcon} />
+          Success!
+        </div>
+      ) : 'Create New Calendar Subscription'}
       subTitle={getSubtitle(isComplete, SUBTITLE)}
       onClose={props.onClose}
       alert={isComplete}
